renamed modalWrapperConfig to modalConfig
