+++ src/app/examples/modal-example/modal-example.component.ts
@@ -1,6 +1,6 @@
 import { Component, ViewContainerRef } from '@angular/core';
 
-import { ModalWrapperConfig } from '~/kirby/components/modal/modal-wrapper/config/modal-wrapper-config';
+import { ModalConfig } from '~/kirby/components/modal/modal-wrapper/config/modal-config';
 import { ModalController } from '~/kirby/components/modal/services/modal.controller';
 import { FirstEmbeddedModalExampleComponent } from './first-embedded-modal-example/first-embedded-modal-example.component';
 
@@ -12,7 +12,7 @@ export class ModalExampleComponent {
   constructor(private modalController: ModalController, private vcRef: ViewContainerRef) {}
 
   showModal() {
-    const config: ModalWrapperConfig = {
+    const config: ModalConfig = {
       title: 'First Embedded Modal',
       titleHorizontalAlignment: 'center',
       component: FirstEmbeddedModalExampleComponent,
