+++ src/kirby/components/modal/modal-wrapper/config/modal-config.helper.ts
@@ -1,23 +1,23 @@
 import { InjectionToken } from '@angular/core';
 
-import { ModalWrapperConfig } from './modal-wrapper-config';
+import { ModalConfig } from './modal-config';
 
-export class ModalWrapperConfigHelper {
+export class ModalConfigHelper {
   static readonly defaultTitleHorizontalAlignment = 'center';
   static readonly defaultCloseIconName = 'close';
   static readonly defaultDim = 0.5;
 
-  static processOptionalValues(config: ModalWrapperConfig): ModalWrapperConfig {
+  static processOptionalValues(config: ModalConfig): ModalConfig {
     if (!config.titleHorizontalAlignment) {
-      config.titleHorizontalAlignment = ModalWrapperConfigHelper.defaultTitleHorizontalAlignment;
+      config.titleHorizontalAlignment = ModalConfigHelper.defaultTitleHorizontalAlignment;
     }
 
     if (!config.closeIconName) {
-      config.closeIconName = ModalWrapperConfigHelper.defaultCloseIconName;
+      config.closeIconName = ModalConfigHelper.defaultCloseIconName;
     }
 
     if (!config.dim) {
-      config.dim = ModalWrapperConfigHelper.defaultDim;
+      config.dim = ModalConfigHelper.defaultDim;
     }
     return config;
   }
