+++ src/kirby/components/modal/modal-wrapper/modal-wrapper.component.ts
@@ -1,8 +1,8 @@
 import { Component, ElementRef, ViewChild, AfterViewInit, Injector } from '@angular/core';
 import { NavParams } from '@ionic/angular';
 
-import { ModalWrapperConfig } from './config/modal-wrapper-config';
-import { COMPONENT_PROPS, ModalWrapperConfigHelper } from './config/modal-wrapper-config.helper';
+import { ModalConfig } from './config/modal-config';
+import { COMPONENT_PROPS, ModalConfigHelper } from './config/modal-config.helper';
 import { IModalController } from '../services/modal.controller.interface';
 
 @Component({
@@ -12,7 +12,7 @@ import { IModalController } from '../services/modal.controller.interface';
 export class ModalWrapperComponent implements AfterViewInit {
   @ViewChild('modalWrapper') modalWrapper: ElementRef;
   scrollY: number = Math.abs(window.scrollY);
-  config: ModalWrapperConfig;
+  config: ModalConfig;
   componentPropsInjector: Injector;
 
   constructor(
@@ -20,7 +20,7 @@ export class ModalWrapperComponent implements AfterViewInit {
     private modalController: IModalController,
     injector: Injector
   ) {
-    this.config = ModalWrapperConfigHelper.processOptionalValues(this.params.get('config'));
+    this.config = ModalConfigHelper.processOptionalValues(this.params.get('config'));
     this.componentPropsInjector = Injector.create({
       providers: [{ provide: COMPONENT_PROPS, useValue: this.params.get('config').componentProps }],
       parent: injector,
