+++ src/kirby/components/modal/services/modal.controller.ts
@@ -4,7 +4,7 @@ import { IModalController } from './modal.controller.interface';
 import { ModalHelper } from './modal.helper';
 import { AlertHelper } from './alert.helper';
 import { ActionSheetHelper } from './action-sheet.helper';
-import { ModalWrapperConfig } from '../modal-wrapper/config/modal-wrapper-config';
+import { ModalConfig } from '../modal-wrapper/config/modal-config';
 import { ActionSheetConfig } from '../action-sheet/config/action-sheet-config';
 import { AlertConfig } from '../alert/config/alert-config';
 
@@ -19,7 +19,7 @@ export class ModalController implements IModalController {
   ) {}
 
   public showModal(
-    config: ModalWrapperConfig,
+    config: ModalConfig,
     vcRef: ViewContainerRef,
     onCloseModal?: (data?: any) => void
   ): void {
