+++ src/kirby/index.ts
@@ -11,6 +11,4 @@ export {
 } from './components/segmented-control/segmented-control.component';
 export { SegmentItem } from './components/segmented-control/segment-item';
 export { BadgeComponent } from './components/badge/badge.component';
-export {
-  COMPONENT_PROPS,
-} from './components/modal/modal-wrapper/config/modal-wrapper-config.helper';
+export { COMPONENT_PROPS } from './components/modal/modal-wrapper/config/modal-config.helper';
